refactor(widgets): admin-layout 폴더명을 admin-sidebar로 변경
설계 문서(PROJECT_PLAN.md)의 widgets/admin-sidebar 명세에 맞게
폴더명 정정 및 하위 모든 파일의 import 경로 업데이트.
