+++ src/pages/admin/missions/index.tsx
@@ -1,4 +1,4 @@
-import { AdminLayout } from '@/widgets/admin-layout/AdminLayout'
+import { AdminLayout } from '@/widgets/admin-sidebar/AdminLayout'
 
 export function AdminMissionsPage() {
   return (
